Относительно большие правки в тестах
- Из-за нового расположения по папкам были проблемы с путями до модулей, исправлено через явное указание пути на директории
- По этой же причины все тесты объеденены
- Все тесты в одном файле __init__.py
